backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.6
numpy==1.24.4
opencv-python==4.8.1.78
pydantic==2.5.2
pydantic-core==2.14.5
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.9.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
midtransclient==1.4.2
pyserial==3.4